requirements/_base.txt
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
coverage==6.4.2
cryptography==37.0.2
cssselect2==0.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
Django==4.0.5
django-admin-honeypot==1.1.0
django-appconf==1.0.5
django-classy-tags==3.0.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-elasticsearch-dsl==7.2.2
django-haystack==3.2.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-json-ld==0.0.5
django-mptt==0.13.4
django-mptt-admin==2.2.0
django-ordered-model==3.6
django-qr-code==3.1.0
django-redis-cache==3.0.1
django-sekizai==3.0.1
django-treebeard==4.5.1
djangorestframework==3.13.1
ecdsa==0.17.0
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.33.3
h11==0.13.0
html5lib==1.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
oauthlib==3.2.0
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pydyf==0.2.0
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyphen==0.12.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-memcached==1.59
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.1
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
segno==1.5.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.3.0
tinycss2==1.1.1
traitlets==5.3.0
transliterate==1.10.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
weasyprint==55.0
webencodings==0.5.1
whitenoise==6.2.0
Whoosh==2.7.4
wrapt==1.14.1
wsproto==1.1.0
zopfli==0.2.1
